Fix stuck scheduler when failed to start probe
Previously, if a probe didn't start we logged the error but still updated
the toplogy. This is bad because if the topology doesn't change we
won't try to start the probe again.
This change mitigate the issue by making sure the current topology
is updated only if all the probes were able to start.
Also made sure the bahavior when a probe is already running is defined:
check if it already starts, if it does do not create a new one.
This ensure that we don't flap all the probes if there is one failing to start