Execute the preparation phase before starting the probe in a coroutine
Before if the prepare phase failed it would start probing (performing check).
If the endpoint is not prepared properly, the main impact will be false
alerts. For example durability check has only meaning if we wrote
the canary data before starting the probe.
After this patch, we make sure we have all prepare checks succeeding before starting the probe.