Disable cgo when building the probe
With newer versions of Go the way the default setting
for cgo_enabled changed and it results in it being enabled.
cgo links the libc which is not practical for deployment on
other system/containers. This commit enforce disabling
of cgo as it was the case in previous build before bumping
Go's version.